# Conflicts:
#	src/main/java/com/example/demo/Controllers/LoginController.java
#	src/main/java/com/example/demo/Repository/UsuarioRepository.java
#	src/main/java/com/example/demo/config/WebSecurityConfig.java

+++ src/main/java/com/example/demo/Controllers/ComunidadController.java
@@ -1,9 +1,12 @@
 package com.example.demo.Controllers;
 
 
+import com.example.demo.Dto.RendirizadorPaginas;
 import com.example.demo.Entity.Comunidad;
 import com.example.demo.Repository.ComunidadRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageRequest;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.validation.BindingResult;
@@ -11,6 +14,7 @@ import org.springframework.web.bind.annotation.*;
 import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
 import javax.validation.Valid;
+import java.awt.print.Pageable;
 import java.util.List;
 import java.util.Optional;
 
@@ -22,7 +26,8 @@ public class ComunidadController {
     ComunidadRepository comunidadRepository;
 
     @GetMapping(value = {"", "/"})
-    public String listaComunidad(Model model){
+    public String listaComunidad( Model model){
+
         model.addAttribute("listaComunidad",comunidadRepository.findAll());
         return "comunidad/lista";
 
@@ -36,14 +41,15 @@ public class ComunidadController {
 
     @GetMapping("/editar")
     public String editarComunidad(@ModelAttribute("comunidad") Comunidad comunidad,
-                                  @RequestParam("id") int id, Model model){
+                                  @RequestParam("id") int id, Model model,RedirectAttributes att){
         Optional<Comunidad> optionalComunidad = comunidadRepository.findById(id);
-        if(optionalComunidad.isPresent()){
+        if(optionalComunidad.isPresent() && optionalComunidad.get().getIdcomunidad()>5 ){
             comunidad = optionalComunidad.get();
             model.addAttribute("comunidad",comunidad);
             return "comunidad/newEdit";
         }
         else {
+            att.addFlashAttribute("msgE", "No puede editar esta Comunidad");
             return  "redirect:/comunidad";
         }
 
@@ -54,12 +60,17 @@ public class ComunidadController {
                                   @RequestParam("id") int id, RedirectAttributes att){
 
         Optional<Comunidad> optionalComunidad = comunidadRepository.findById(id);
-        if(optionalComunidad.isPresent()){
+        if(optionalComunidad.isPresent() && optionalComunidad.get().getIdcomunidad()>5){
             att.addFlashAttribute("msgCo", "Comunidad Borrada Exitosamente");
             comunidadRepository.deleteById(id);
             return "redirect:/comunidad";
         }
-        return "redirect:/comunidad";
+        else
+        {
+            att.addFlashAttribute("msgE", "No puede borrar esta Comunidad");
+            return "redirect:/comunidad";
+        }
+
     }
 
     @PostMapping("/guardar")
@@ -72,11 +83,11 @@ public class ComunidadController {
             if(comunidad.getIdcomunidad()==0){
                 for (Comunidad com : comunidadRepository.findAll()) {
                     if (com.getNombrecomunidad().equalsIgnoreCase(comunidad.getNombrecomunidad()) || com.getCodigocomunidad().equalsIgnoreCase(comunidad.getCodigocomunidad())) {
-                        if (com.getNombrecomunidad().equals(comunidad.getNombrecomunidad())) {
+                        if (com.getNombrecomunidad().equalsIgnoreCase(comunidad.getNombrecomunidad())) {
                             att.addFlashAttribute("msg1", "Nombre de Comunidad ya exite");
                             att.addFlashAttribute("comunidad", comunidad);
                         }
-                        if (com.getCodigocomunidad().equals(comunidad.getCodigocomunidad())) {
+                        if (com.getCodigocomunidad().equalsIgnoreCase(comunidad.getCodigocomunidad())) {
                             att.addFlashAttribute("msg2", "Codigo de Comunidad ya existe");
                             att.addFlashAttribute("comunidad", comunidad);
                         }
@@ -92,11 +103,11 @@ public class ComunidadController {
             else{
                 for(Comunidad comunidad1:comunidadRepository.mio(comunidad.getIdcomunidad())){
                     if (comunidad1.getNombrecomunidad().equalsIgnoreCase(comunidad.getNombrecomunidad()) || comunidad1.getCodigocomunidad().equalsIgnoreCase(comunidad.getCodigocomunidad())) {
-                        if (comunidad1.getNombrecomunidad().equals(comunidad.getNombrecomunidad())) {
+                        if (comunidad1.getNombrecomunidad().equalsIgnoreCase(comunidad.getNombrecomunidad())) {
                             att.addFlashAttribute("msg1", "Nombre de Comunidad ya exite");
                             att.addFlashAttribute("comunidad", comunidad);
                         }
-                        if (comunidad1.getCodigocomunidad().equals(comunidad.getCodigocomunidad())) {
+                        if (comunidad1.getCodigocomunidad().equalsIgnoreCase(comunidad.getCodigocomunidad())) {
                             att.addFlashAttribute("msg2", "Codigo de Comunidad ya existe");
                             att.addFlashAttribute("comunidad", comunidad);
                         }
@@ -109,6 +120,9 @@ public class ComunidadController {
                     }
                 }
             }
+            //String nom= comunidad.getNombrecomunidad().substring(0, 1).toUpperCase() + comunidad.getNombrecomunidad().substring(1);
+            String nom=comunidad.getNombrecomunidad().substring(0, 1).toUpperCase() + comunidad.getNombrecomunidad().substring(1).toLowerCase();
+            comunidad.setNombrecomunidad(nom);
             String cod=comunidad.getCodigocomunidad().toUpperCase();
             comunidad.setCodigocomunidad(cod);
             comunidadRepository.save(comunidad);
@@ -117,9 +131,17 @@ public class ComunidadController {
     }
 
     @PostMapping("/buscar")
-    public String filtar(@RequestParam("nom") String nom,Model model){
-        model.addAttribute("listaComunidad",comunidadRepository.filtro(nom));
-        return "comunidad/lista";
+    public String filtar(@RequestParam("nom") String nom,Model model,RedirectAttributes att){
+        if(nom.isEmpty()){
+            att.addFlashAttribute("msgBuscador", "Campo vacio. Ingrese el dato a buscar");
+            return "redirect:/comunidad";
+        }
+        else{
+
+            model.addAttribute("listaComunidad",comunidadRepository.filtro(nom));
+            return "comunidad/lista";
+        }
+
     }
